requirements.txt
selenium
pillow
pytesseract
fake-useragent